+++ packages/client/src/app/pages/my-stuff/my-stuff.component.html
@@ -10,7 +10,7 @@
                 <button *ngSwitchCase="pubStatus.Published" matRipple class="tool active">
                     <i-feather name="check-circle"></i-feather>Published
                 </button>
-                <button *ngSwitchCase="pubStatus.Unpublished" matRipple class="tool" (click)="publishContent(currContent)">
+                <button *ngSwitchCase="pubStatus.Unpublished" matRipple class="tool">
                     <i-feather name="circle"></i-feather>Unpublished
                 </button>
                 <button *ngSwitchCase="pubStatus.Pending" matRipple class="tool">
@@ -26,19 +26,19 @@
             <span class="tool-divider"></span>
             
             <!--Editing-->
-            <button matRipple class="tool" (click)="viewContent(currContent)">
+            <button matRipple class="tool">
                 <i-feather name="edit-3"></i-feather>Edit
             </button>
             <span class="tool-divider"></span>
 
             <!--General Tools-->
-            <button matRipple class="tool left" (click)="deselect()" matTooltip="Deselect" matTooltipClass="offprint-tooltip">
+            <button matRipple class="tool left" matTooltip="Deselect" matTooltipClass="offprint-tooltip">
                 <i-feather class="no-text" name="x"></i-feather>
             </button>
             <button matRipple class="tool mid" [matTooltip]="'Share'" [matTooltipClass]="'offprint-tooltip'" [matMenuTriggerFor]="shareMenu" [matMenuTriggerData]="{content: currContent}">
                 <i-feather class="no-text" name="share"></i-feather>
             </button>
-            <button matRipple class="tool end" (click)="deleteContent(currContent)" matTooltip="Delete" matTooltipClass="offprint-tooltip">
+            <button matRipple class="tool end" matTooltip="Delete" matTooltipClass="offprint-tooltip">
                 <i-feather class="no-text" name="trash-2"></i-feather>
             </button>
             <span class="tool-divider"></span>
@@ -105,7 +105,7 @@
                 <button mat-menu-item *ngSwitchCase="contentKind.PoetryContent" [routerLink]="['/poetry', content._id, content.title | slugify]">Go to Public Page</button>
             </ng-container>
         </ng-container>
-        <button mat-menu-item (click)="getShareLink(content)">Copy Share Link</button>
+        <button mat-menu-item>Copy Share Link</button>
     </ng-template>
 </mat-menu>
 
@@ -149,7 +149,7 @@
             <div class="content-block">
                 <div class="content-grid">
                     <ng-container *ngFor="let content of myContent">
-                        <content-item [content]="content" [selected]="itemSelected" (selectItem)="selectItem($event)" (viewItem)="viewContent($event)"></content-item>
+                        <content-item [content]="content"></content-item>
                     </ng-container>
                 </div>
             </div>
